requirements.txt
telethon==1.27.0
python-dotenv==1.1.1